Finished mmpbsa. Tested the program and the results match those of the python version of mmpbsa.

+++ mmpbsa_utils.h
@@ -25,7 +25,7 @@
 //MEAD
 #include "MEAD/Coord.h"
 
-typedef float mmpbsa_t;
+typedef double mmpbsa_t;
 extern std::fstream myOutput;
